+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Fixed
+
+- Exposed `conflate` through the public socket option accessors.
+
 ## [0.28.3] - 2026-07-30
 
 ### Added